+++ config/config.exs
@@ -181,14 +181,12 @@ config :tesla, :adapter, {Tesla.Adapter.Finch, name: MyFinch}
 
 # Configures http settings, upstream proxy etc.
 config :pleroma, :http,
-  pool_timeout: :timer.seconds(5),
+  pool_timeout: :timer.seconds(60),
   receive_timeout: :timer.seconds(15),
   proxy_url: nil,
   user_agent: :default,
   pool_size: 10,
-  adapter: [],
-  # see: https://hexdocs.pm/finch/Finch.html#start_link/1
-  pool_max_idle_time: :timer.seconds(30)
+  adapter: []
 
 config :pleroma, :instance,
   name: "Akkoma",